refactor: defer loading of routed modals (#11632)
