--- packages/mask/popups/modals/modals.ts
@@ -1,16 +1,0 @@
-export * from './modal-controls.js'
-
-export * from './ChooseCurrencyModal/index.js'
-export * from './ChooseNetworkModal/index.js'
-export * from './ConnectSocialAccountModal/index.js'
-export * from './SelectProviderModal/index.js'
-export * from './SwitchPersonaModal/index.js'
-export * from './PersonaSettingModal/index.js'
-export * from './PersonaRenameModal/index.js'
-export * from './SetBackupPasswordModal/index.js'
-export * from './VerifyBackupPasswordModal/index.js'
-export * from './WalletGroupModal/index.js'
-export * from './SelectLanguageModal/index.js'
-export * from './SelectAppearanceModal/index.js'
-export * from './SupportedSitesModal/index.js'
-export * from './ChangeBackupPasswordModal/index.js'
