+++ packages/mask/popups/pages/Wallet/CollectibleDetail/index.tsx
@@ -19,7 +19,7 @@ import { useLocation, useNavigate } from 'react-router-dom'
 import urlcat from 'urlcat'
 import { useMaskSharedTrans } from '../../../../shared-ui/index.js'
 import { PageTitleContext, useTitle, useTokenParams } from '../../../hooks/index.js'
-import { ConfirmModal } from '../../../modals/modals.js'
+import { ConfirmModal } from '../../../modals/modal-controls.js'
 import { TransferTabType } from '../type.js'
 
 const useStyles = makeStyles()((theme) => ({
