+++ packages/mask/popups/pages/Wallet/TokenDetail/index.tsx
@@ -23,7 +23,7 @@ import React, { memo, useContext, useEffect, useMemo, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useMaskSharedTrans } from '../../../../shared-ui/index.js'
 import { PageTitleContext, useTitle, useTokenParams } from '../../../hooks/index.js'
-import { ConfirmModal } from '../../../modals/modals.js'
+import { ConfirmModal } from '../../../modals/modal-controls.js'
 import { ActionGroup } from '../components/index.js'
 import { useAsset } from '../hooks/index.js'
 import { DIMENSION, TrendingChart } from './TrendingChart.js'
