+++ packages/mask/popups/pages/Wallet/WalletSettings/ChangePaymentPassword.tsx
@@ -2,7 +2,7 @@ import { Icons } from '@masknet/icons'
 import { Box, ListItem, Typography } from '@mui/material'
 import { useMaskSharedTrans } from '../../../../shared-ui/index.js'
 import { useStyles } from './useStyles.js'
-import { ChangePaymentPasswordModal } from '../../../modals/modals.js'
+import { ChangePaymentPasswordModal } from '../../../modals/modal-controls.js'
 
 export function ChangePaymentPassword() {
     const t = useMaskSharedTrans()
