+++ packages/mask/popups/pages/Wallet/WalletSettings/ShowPrivateKey.tsx
@@ -2,7 +2,7 @@ import { Icons } from '@masknet/icons'
 import { Box, ListItem, Typography } from '@mui/material'
 import { useMaskSharedTrans } from '../../../../shared-ui/index.js'
 import { useStyles } from './useStyles.js'
-import { ShowPrivateKeyModal } from '../../../modals/modals.js'
+import { ShowPrivateKeyModal } from '../../../modals/modal-controls.js'
 
 export function ShowPrivateKey() {
     const t = useMaskSharedTrans()
