+++ packages/mask/popups/pages/Wallet/WalletSettings/index.tsx
@@ -12,7 +12,7 @@ import { Trans } from 'react-i18next'
 import { useMaskSharedTrans } from '../../../../shared-ui/index.js'
 import { useModalNavigate } from '../../../components/index.js'
 import { useTitle } from '../../../hooks/index.js'
-import { WalletRemoveModal } from '../../../modals/modals.js'
+import { WalletRemoveModal } from '../../../modals/modal-controls.js'
 import { AutoLock } from './AutoLock.js'
 import { ChangeCurrency } from './ChangeCurrency.js'
 import { ChangeNetwork } from './ChangeNetwork.js'
